Refactoring the playground into a module to drop use of globals
