+++ examples/index.html
@@ -5,7 +5,7 @@
     <!-- Add the A-Frame and Smooth Voxels scripts -->
     <!-- NOTE: Do NOT add a script tag for smoothvoxelworker.js. It will load automatically. -->
     <script src="https://aframe.io/releases/1.3.0/aframe.min.js"></script>
-    <script src="dist/smoothvoxels.js"></script>
+    <script src="dist/svox.js"></script>
     
     <script>   
 SVOX = { models: {} };
