+++ package.json
@@ -28,5 +28,8 @@
   },
   "peerDependencies": {
     "aframe": "^1.3.0"
+  },
+  "dependencies": {
+    "buffer": "^6.0.3"
   }
 }
