+++ src/parse-magica-voxel/PACK.js
@@ -1,0 +1,3 @@
+export default function PACKHandler (state) {
+  return state.Buffer.readInt32LE(state.readByteIndex)
+};
