+++ src/parse-magica-voxel/SKIP.js
@@ -1,0 +1,5 @@
+// Skip this chunk.
+export default function SKIPHandler (state, startIndex, endIndex) {
+  state.readByteIndex = endIndex
+  return { error: 'Unsupported chunk type' }
+}
