+++ src/parse-magica-voxel/XYZI.js
@@ -1,0 +1,17 @@
+export default function XYZIHandler (state, startIndex, endIndex) {
+  const numVoxels = Math.abs(state.Buffer.readInt32LE(state.readByteIndex))
+  state.readByteIndex += 4
+
+  const voxelData = []
+  for (let n = 0; n < numVoxels; n++) {
+    voxelData[n] = {
+      x: state.Buffer[state.readByteIndex++] & 0xFF,
+      y: state.Buffer[state.readByteIndex++] & 0xFF,
+      z: state.Buffer[state.readByteIndex++] & 0xFF,
+      c: state.Buffer[state.readByteIndex++] & 0xFF // color index in RGBA
+    }
+  }
+
+  console.assert(state.readByteIndex === endIndex, 'XYZI chunk did not fully read')
+  return voxelData
+};
