+++ src/parse-magica-voxel/constants.js
@@ -1,0 +1,1 @@
+export const intByteLength = 4
