+++ src/parse-magica-voxel/readId.js
@@ -1,0 +1,8 @@
+export default function readId (state) {
+  const id = String.fromCharCode(parseInt(state.Buffer[state.readByteIndex++])) +
+           String.fromCharCode(parseInt(state.Buffer[state.readByteIndex++])) +
+           String.fromCharCode(parseInt(state.Buffer[state.readByteIndex++])) +
+           String.fromCharCode(parseInt(state.Buffer[state.readByteIndex++]))
+
+  return id
+};
